feat(gofish): consolidate settings dialog and inline hint toggle into SettingsPanel
GoFishPage used a one-off GoFishSettingsDialog wrapper (difficulty select
only) plus a separate inline hint checkbox floating next to the deck
counter. Fold both into a single direct <SettingsPanel> with the
difficulty select and hint toggle as items, and delete the now-dead
GoFishSettingsDialog component and its test.

Closes #2202
